Refactor functional tests into new package
