+++ src/e2e/java/uk/gov/hmcts/reform/sscs/functional/sya/notifications/AppealDormantNotificationTest.java
@@ -1,10 +1,11 @@
-package uk.gov.hmcts.reform.sscs.functional;
+package uk.gov.hmcts.reform.sscs.functional.sya.notifications;
 
 import static uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType.APPEAL_DORMANT_NOTIFICATION;
 
 import java.util.List;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Value;
+import uk.gov.hmcts.reform.sscs.functional.AbstractFunctionalTest;
 import uk.gov.service.notify.Notification;
 
 public class AppealDormantNotificationTest extends AbstractFunctionalTest {
